requirements.txt
asgiref==3.9.1
Django==5.2.5
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
# psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-allauth==65.3.1
gunicorn
whitenoise
